students management

+++ db/migrate/20170403152555_create_subjects.rb
@@ -3,7 +3,7 @@ class CreateSubjects < ActiveRecord::Migration[5.0]
     create_table :subjects do |t|
       t.string :subject_id
       t.string :name
-      t.integer :num_of_credit
+      t.float :num_of_credit
 
       t.timestamps
     end
